rename package org.oskari.wcs.util.smallmap to org.oskari.wcs.util.small
