+++ service-wcs/src/main/java/org/oskari/wcs/request/GetCapabilities.java
@@ -1,7 +1,8 @@
 package org.oskari.wcs.request;
 
+import org.oskari.wcs.util.small.SmallMap;
+
 import java.util.Map;
-import org.oskari.wcs.util.smallmap.SmallMap;
 
 public class GetCapabilities {
