+++ service-wcs/src/main/java/org/oskari/wcs/util/small/SmallEntry.java
@@ -1,4 +1,4 @@
-package org.oskari.wcs.util.smallmap;
+package org.oskari.wcs.util.small;
 
 import java.util.Map;
 import java.util.Objects;
