+++ service-wcs/src/main/java/org/oskari/wcs/util/small/SmallMap.java
@@ -1,4 +1,4 @@
-package org.oskari.wcs.util.smallmap;
+package org.oskari.wcs.util.small;
 
 import java.util.Arrays;
 import java.util.Collection;
